requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pillow
torch
transformers
